# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ISAGI NEWS API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://isagiapi.galder315.ga/"
    },
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/NewUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario registrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Usuario ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario logueado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales no validas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/request-recover": {
      "post": {
        "tags": ["auth"],
        "description": "Solicita el envio de un correo electronico para recuperar la contraseña",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RecoverRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Correo enviado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/recover": {
      "post": {
        "tags": ["auth"],
        "description": "Recupera la contraseña de un usuario",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RecoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contraseña recuperada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos no validos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MessageResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "message": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "error": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        }
      },
      "NewUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          }
        },
        "required": ["email", "password"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          },
          "remember": {
            "type": "boolean",
            "description": "Recordar usuario"
          }
        },
        "required": ["email", "password", "remember"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Codigo de estado de la respuesta"
          },
          "token": {
            "type": "string",
            "description": "Token de autenticacion"
          }
        }
      },
      "RecoverRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "token": {
            "type": "string",
            "description": "Token de recuperacion de contraseña"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          }
        },
        "required": ["email", "token", "password"]
      },
      "RecoverRequestRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario"
          }
        },
        "required": ["email"]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}